+++ CHANGELOG.rst
@@ -10,6 +10,9 @@ CHANGELOG
   would result in a broken markup after upgrading
 * Fixed a bug where whitespace would be incorrectly removed before the child
   plugin on save of the text plugin
+* Changed the way ckeditor js is loaded, everything is now loaded in one bundle
+* Changed the way ckeditor loads must-have external plugins, so user config
+  would have no effect on the settings
 
 
 3.1.0 (2016-08-18)